pulls out routes and header into separate components
